+++ src/Docker.PowerShell/Cmdlets/RemoveContainerNetwork.cs
@@ -4,6 +4,7 @@ using System.Threading.Tasks;
 namespace Docker.PowerShell.Cmdlets;
 
 [Cmdlet(VerbsCommon.Remove, "ContainerNet",
+        SupportsShouldProcess = true,
         DefaultParameterSetName = CommonParameterSetNames.Default)]
 public class RemoveContainerNet : NetworkOperationCmdlet
 {
@@ -23,6 +24,11 @@ public class RemoveContainerNet : NetworkOperationCmdlet
     {
         foreach (var id in ParameterResolvers.GetNetworkIds(Network, Id))
         {
+            if (!ShouldProcess(id, "Remove network"))
+            {
+                continue;
+            }
+
             await DkrClient.Networks.DeleteNetworkAsync(id);
         }
     }
